Fixed test failure after rebase * Removed unused imports and redundant throws declarations
