+++ src/main/java/org/schemaspy/Config.java
@@ -29,8 +29,6 @@ import org.schemaspy.view.DefaultSqlFormatter;
 import org.schemaspy.view.SqlFormatter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.slf4j.event.Level;
-
 
 import java.beans.BeanInfo;
 import java.beans.IntrospectionException;
