+++ src/main/java/org/schemaspy/DbDriverLoader.java
@@ -22,7 +22,6 @@ import org.schemaspy.model.ConnectionFailure;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.lang.invoke.MethodHandles;
 import java.net.MalformedURLException;
