+++ src/test/java/org/schemaspy/app/integrationtesting/H2SpacesIT.java
@@ -1,4 +1,4 @@
-package org.schemaspy.integrationtesting;
+package org.schemaspy.app.integrationtesting;
 
 import org.junit.Before;
 import org.junit.ClassRule;
@@ -6,7 +6,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.schemaspy.Config;
-import org.schemaspy.cli.CommandLineArguments;
+import org.schemaspy.app.cli.CommandLineArguments;
 import org.schemaspy.model.Database;
 import org.schemaspy.model.ProgressListener;
 import org.schemaspy.service.DatabaseService;
