+++ src/test/java/org/schemaspy/app/integrationtesting/MysqlKeyWordTable.java
@@ -1,4 +1,4 @@
-package org.schemaspy.testcontainer;
+package org.schemaspy.app.integrationtesting;
 
 import com.github.npetzall.testcontainers.junit.jdbc.JdbcContainerRule;
 import org.junit.Before;
@@ -7,8 +7,9 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.schemaspy.Config;
-import org.schemaspy.cli.CommandLineArguments;
-import org.schemaspy.model.*;
+import org.schemaspy.app.cli.CommandLineArguments;
+import org.schemaspy.model.Database;
+import org.schemaspy.model.ProgressListener;
 import org.schemaspy.service.DatabaseService;
 import org.schemaspy.service.SqlService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,7 +32,7 @@ import static org.mockito.BDDMockito.given;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class MysqlSpacesNoDotsIT {
+public class MysqlKeyWordTable {
 
     @Autowired
     private SqlService sqlService;
@@ -55,8 +56,7 @@ public class MysqlSpacesNoDotsIT {
             new JdbcContainerRule<>(() -> new MySQLContainer())
                     .assumeDockerIsPresent()
                     .withAssumptions(assumeDriverIsPresent())
-                    .withInitScript("integrationTesting/dbScripts/mysql_spaces_no_dots.sql")
-                    .withInitUser("root", "test");
+                    .withInitScript("integrationTesting/dbScripts/mysql_keyword.sql");
 
     @Before
     public synchronized void createDatabaseRepresentation() throws SQLException, IOException, ScriptException, URISyntaxException {
@@ -68,21 +68,21 @@ public class MysqlSpacesNoDotsIT {
     private void doCreateDatabaseRepresentation() throws SQLException, IOException, URISyntaxException {
         String[] args = {
                 "-t", "mysql",
-                "-db", "TEST 1",
-                "-s", "TEST 1",
+                "-db", "test",
+                "-s", "test",
                 "-cat", "%",
-                "-o", "target/integrationtesting/mysql_spaces_no_dots",
+                "-o", "target/integrationtesting/mysql_keywords",
                 "-u", "test",
                 "-p", "test",
                 "-host", jdbcContainerRule.getContainer().getContainerIpAddress(),
                 "-port", jdbcContainerRule.getContainer().getMappedPort(3306).toString()
         };
-        given(arguments.getOutputDirectory()).willReturn(new File("target/integrationtesting/mysql_spaces_no_dots"));
+        given(arguments.getOutputDirectory()).willReturn(new File("target/integrationtesting/mysql_keywords"));
         given(arguments.getDatabaseType()).willReturn("mysql");
         given(arguments.getUser()).willReturn("test");
-        given(arguments.getSchema()).willReturn("TEST 1");
+        given(arguments.getSchema()).willReturn("test");
         given(arguments.getCatalog()).willReturn("%");
-        given(arguments.getDatabaseName()).willReturn("TEST 1");
+        given(arguments.getDatabaseName()).willReturn("test");
         Config config = new Config(args);
         DatabaseMetaData databaseMetaData = sqlService.connect(config);
         Database database = new Database(config, databaseMetaData, arguments.getDatabaseName(), arguments.getCatalog(), arguments.getSchema(), null, progressListener);
@@ -91,35 +91,7 @@ public class MysqlSpacesNoDotsIT {
     }
 
     @Test
-    public void databaseShouldExist() {
-        assertThat(database).isNotNull();
-        assertThat(database.getName()).isEqualToIgnoringCase("TEST 1");
-    }
-
-    @Test
-    public void databaseShouldHaveTable() {
-        assertThat(database.getTables()).extracting(Table::getName).contains("TABLE 1");
-    }
-
-    @Test
-    public void tableShouldHavePKWithAutoIncrement() {
-        assertThat(database.getTablesByName().get("TABLE 1").getColumns()).extracting(TableColumn::getName).contains("id");
-        assertThat(database.getTablesByName().get("TABLE 1").getColumn("id").isPrimary()).isTrue();
-        assertThat(database.getTablesByName().get("TABLE 1").getColumn("id").isAutoUpdated()).isTrue();
-    }
-
-    @Test
-    public void tableShouldHaveForeignKey() {
-        assertThat(database.getTablesByName().get("TABLE 1").getForeignKeys()).extracting(ForeignKeyConstraint::getName).contains("link fk");
-    }
-
-    @Test
-    public void tableShouldHaveUniqueKey() {
-        assertThat(database.getTablesByName().get("TABLE 1").getIndexes()).extracting(TableIndex::getName).contains("name_link_unique");
-    }
-
-    @Test
-    public void tableShouldHaveColumnWithSpaceInIt() {
-        assertThat(database.getTablesByName().get("TABLE 1").getColumns()).extracting(TableColumn::getName).contains("link id");
+    public void hasATableNamedDistinct() {
+        assertThat(database.getTables()).extracting(t -> t.getName()).contains("DISTINCT");
     }
 }
